refactor(web): StatusScreen/MutedLink components, per-icon files, cva Button
pageStyles.ts exported raw className string constants — the kind of
non-component reuse the styling ADR already rules out, since the React
component (not a CSS class) is meant to be the reuse boundary. Replaced with
StatusScreen and MutedLink components. icons.tsx splits into one file per
icon plus an index.ts barrel for maintainability. Button's hand-rolled
variant/size Record maps become a class-variance-authority config.
