--- apps/web/src/components/icons.tsx
@@ -1,80 +1,0 @@
-import type { ReactNode } from "react";
-
-interface IconProps {
-  size?: number;
-}
-
-function Svg({ size = 14, children }: IconProps & { children: ReactNode }) {
-  return (
-    <svg
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth={1.5}
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      width={size}
-      height={size}
-      aria-hidden="true"
-    >
-      {children}
-    </svg>
-  );
-}
-
-export function CheckIcon(props: IconProps) {
-  return (
-    <Svg {...props}>
-      <path d="m4.5 12.75 6 6 9-13.5" />
-    </Svg>
-  );
-}
-
-export function LockIcon(props: IconProps) {
-  return (
-    <Svg {...props}>
-      <path d="M16.5 10.5V6.75a4.5 4.5 0 1 0-9 0v3.75m-.75 11.25h10.5a2.25 2.25 0 0 0 2.25-2.25v-6.75a2.25 2.25 0 0 0-2.25-2.25H6.75a2.25 2.25 0 0 0-2.25 2.25v6.75a2.25 2.25 0 0 0 2.25 2.25Z" />
-    </Svg>
-  );
-}
-
-export function ChevronRightIcon(props: IconProps) {
-  return (
-    <Svg {...props}>
-      <path d="m8.25 4.5 7.5 7.5-7.5 7.5" />
-    </Svg>
-  );
-}
-
-export function PlayIcon(props: IconProps) {
-  return (
-    <Svg {...props}>
-      <path d="M5.25 5.653c0-.856.917-1.398 1.667-.986l11.54 6.348a1.125 1.125 0 0 1 0 1.971l-11.54 6.347a1.125 1.125 0 0 1-1.667-.985V5.653Z" />
-    </Svg>
-  );
-}
-
-export function CompassIcon(props: IconProps) {
-  return (
-    <Svg {...props}>
-      <path d="M15 10.5a3 3 0 1 1-6 0 3 3 0 0 1 6 0Z" />
-      <path d="M19.5 10.5c0 7.142-7.5 11.25-7.5 11.25S4.5 17.642 4.5 10.5a7.5 7.5 0 1 1 15 0Z" />
-    </Svg>
-  );
-}
-
-export function SearchIcon(props: IconProps) {
-  return (
-    <Svg {...props}>
-      <path d="m21 21-5.197-5.197m0 0A7.5 7.5 0 1 0 5.196 5.196a7.5 7.5 0 0 0 10.607 10.607Z" />
-    </Svg>
-  );
-}
-
-export function InfoIcon(props: IconProps) {
-  return (
-    <Svg {...props}>
-      <path d="m11.25 11.25.041-.02a.75.75 0 0 1 1.063.852l-.708 2.836a.75.75 0 0 0 1.063.853l.041-.021M21 12a9 9 0 1 1-18 0 9 9 0 0 1 18 0Zm-9-3.75h.008v.008H12V8.25Z" />
-    </Svg>
-  );
-}
