+++ apps/web/src/components/icons/CheckIcon.tsx
@@ -1,0 +1,9 @@
+import { Svg, type IconProps } from "./Svg";
+
+export function CheckIcon(props: IconProps) {
+  return (
+    <Svg {...props}>
+      <path d="m4.5 12.75 6 6 9-13.5" />
+    </Svg>
+  );
+}
