+++ apps/web/src/components/icons/ChevronRightIcon.tsx
@@ -1,0 +1,9 @@
+import { Svg, type IconProps } from "./Svg";
+
+export function ChevronRightIcon(props: IconProps) {
+  return (
+    <Svg {...props}>
+      <path d="m8.25 4.5 7.5 7.5-7.5 7.5" />
+    </Svg>
+  );
+}
