+++ apps/web/src/components/icons/CompassIcon.tsx
@@ -1,0 +1,10 @@
+import { Svg, type IconProps } from "./Svg";
+
+export function CompassIcon(props: IconProps) {
+  return (
+    <Svg {...props}>
+      <path d="M15 10.5a3 3 0 1 1-6 0 3 3 0 0 1 6 0Z" />
+      <path d="M19.5 10.5c0 7.142-7.5 11.25-7.5 11.25S4.5 17.642 4.5 10.5a7.5 7.5 0 1 1 15 0Z" />
+    </Svg>
+  );
+}
