+++ apps/web/src/components/icons/LockIcon.tsx
@@ -1,0 +1,9 @@
+import { Svg, type IconProps } from "./Svg";
+
+export function LockIcon(props: IconProps) {
+  return (
+    <Svg {...props}>
+      <path d="M16.5 10.5V6.75a4.5 4.5 0 1 0-9 0v3.75m-.75 11.25h10.5a2.25 2.25 0 0 0 2.25-2.25v-6.75a2.25 2.25 0 0 0-2.25-2.25H6.75a2.25 2.25 0 0 0-2.25 2.25v6.75a2.25 2.25 0 0 0 2.25 2.25Z" />
+    </Svg>
+  );
+}
