+++ apps/web/src/components/icons/PlayIcon.tsx
@@ -1,0 +1,9 @@
+import { Svg, type IconProps } from "./Svg";
+
+export function PlayIcon(props: IconProps) {
+  return (
+    <Svg {...props}>
+      <path d="M5.25 5.653c0-.856.917-1.398 1.667-.986l11.54 6.348a1.125 1.125 0 0 1 0 1.971l-11.54 6.347a1.125 1.125 0 0 1-1.667-.985V5.653Z" />
+    </Svg>
+  );
+}
