+++ apps/web/src/components/icons/SearchIcon.tsx
@@ -1,0 +1,9 @@
+import { Svg, type IconProps } from "./Svg";
+
+export function SearchIcon(props: IconProps) {
+  return (
+    <Svg {...props}>
+      <path d="m21 21-5.197-5.197m0 0A7.5 7.5 0 1 0 5.196 5.196a7.5 7.5 0 0 0 10.607 10.607Z" />
+    </Svg>
+  );
+}
