+++ apps/web/src/components/icons/index.ts
@@ -1,0 +1,8 @@
+export type { IconProps } from "./Svg";
+export { CheckIcon } from "./CheckIcon";
+export { LockIcon } from "./LockIcon";
+export { ChevronRightIcon } from "./ChevronRightIcon";
+export { PlayIcon } from "./PlayIcon";
+export { CompassIcon } from "./CompassIcon";
+export { SearchIcon } from "./SearchIcon";
+export { InfoIcon } from "./InfoIcon";
