--- apps/web/src/lib/pageStyles.ts
@@ -1,3 +1,0 @@
-export const STATUS_SCREEN_CLASSES = "flex min-h-[100svh] flex-col px-0 py-10 text-center";
-
-export const LINK_MUTED_CLASSES = "mt-4 block text-center text-[13px] text-accent no-underline";
